+++ packages/admin/cms-admin/src/blocks/helpers/composeBlocks/createCompositeBlockFields.tsx
@@ -1,9 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-<<<<<<< HEAD:packages/admin/cms-admin/src/blocks/helpers/composeBlocks/createCompositeBlockFields.tsx
-import { type AnonymousBlockInterface, type BlockAdminComponent } from "../../types";
-=======
-import { AnonymousBlockInterface, BlockAdminComponent, BlockMethods } from "../../types";
->>>>>>> main:packages/admin/blocks-admin/src/blocks/helpers/composeBlocks/createCompositeSettings.tsx
+import { type AnonymousBlockInterface, type BlockAdminComponent, type BlockMethods } from "../../types";
 import { createSettingsAnonymousBlock } from "./createSettingsBlock";
 
 interface Options<State extends Record<string, any>> {